+++ cmd/src/login_test.go
@@ -9,6 +9,8 @@ import (
 	"net/http/httptest"
 	"strings"
 	"testing"
+
+	"github.com/sourcegraph/src-cli/internal/cmderrors"
 )
 
 func TestLogin(t *testing.T) {
@@ -22,7 +24,7 @@ func TestLogin(t *testing.T) {
 
 	t.Run("different endpoint in config vs. arg", func(t *testing.T) {
 		out, err := check(t, &config{Endpoint: "https://example.com"}, "https://sourcegraph.example.com")
-		if err != exitCode1 {
+		if err != cmderrors.ExitCode1 {
 			t.Fatal(err)
 		}
 		wantOut := "❌ Problem: No access token is configured.\n\n🛠  To fix: Create an access token at https://sourcegraph.example.com/user/settings/tokens, then set the following environment variables:\n\n   SRC_ENDPOINT=https://sourcegraph.example.com\n   SRC_ACCESS_TOKEN=(the access token you just created)\n\n   To verify that it's working, run this command again."
@@ -33,7 +35,7 @@ func TestLogin(t *testing.T) {
 
 	t.Run("no access token", func(t *testing.T) {
 		out, err := check(t, &config{Endpoint: "https://example.com"}, "https://sourcegraph.example.com")
-		if err != exitCode1 {
+		if err != cmderrors.ExitCode1 {
 			t.Fatal(err)
 		}
 		wantOut := "❌ Problem: No access token is configured.\n\n🛠  To fix: Create an access token at https://sourcegraph.example.com/user/settings/tokens, then set the following environment variables:\n\n   SRC_ENDPOINT=https://sourcegraph.example.com\n   SRC_ACCESS_TOKEN=(the access token you just created)\n\n   To verify that it's working, run this command again."
@@ -44,7 +46,7 @@ func TestLogin(t *testing.T) {
 
 	t.Run("warning when using config file", func(t *testing.T) {
 		out, err := check(t, &config{Endpoint: "https://example.com", ConfigFilePath: "f"}, "https://example.com")
-		if err != exitCode1 {
+		if err != cmderrors.ExitCode1 {
 			t.Fatal(err)
 		}
 		wantOut := "⚠️  Warning: Configuring src with a JSON file is deprecated. Please migrate to using the env vars SRC_ENDPOINT and SRC_ACCESS_TOKEN instead, and then remove f. See https://github.com/sourcegraph/src-cli#readme for more information.\n\n❌ Problem: No access token is configured.\n\n🛠  To fix: Create an access token at https://example.com/user/settings/tokens, then set the following environment variables:\n\n   SRC_ENDPOINT=https://example.com\n   SRC_ACCESS_TOKEN=(the access token you just created)\n\n   To verify that it's working, run this command again."
@@ -62,7 +64,7 @@ func TestLogin(t *testing.T) {
 
 		endpoint := s.URL
 		out, err := check(t, &config{Endpoint: endpoint, AccessToken: "x"}, endpoint)
-		if err != exitCode1 {
+		if err != cmderrors.ExitCode1 {
 			t.Fatal(err)
 		}
 		wantOut := "❌ Problem: Invalid access token.\n\n🛠  To fix: Create an access token at $ENDPOINT/user/settings/tokens, then set the following environment variables:\n\n   SRC_ENDPOINT=$ENDPOINT\n   SRC_ACCESS_TOKEN=(the access token you just created)\n\n   To verify that it's working, run this command again.\n\n   (If you need to supply custom HTTP request headers, see information about SRC_HEADER_* env vars at https://github.com/sourcegraph/src-cli/blob/main/AUTH_PROXY.md.)"